chore: Refactor prisma for threat_level_code
