+++ backend/src/prisma/cm/prisma.cm.service.ts
@@ -1,0 +1,13 @@
+import { Injectable, OnModuleInit, OnModuleDestroy } from "@nestjs/common";
+import { PrismaClient } from "../../../prisma/case_management/generated";
+
+@Injectable()
+export class CaseManagementPrismaService extends PrismaClient implements OnModuleInit, OnModuleDestroy {
+  async onModuleInit() {
+    await this.$connect();
+  }
+
+  async onModuleDestroy() {
+    await this.$disconnect();
+  }
+}
