--- backend/src/prisma/prisma.module.ts
@@ -1,8 +1,0 @@
-import { Module } from "@nestjs/common";
-import { PrismaService } from "nestjs-prisma";
-
-@Module({
-  providers: [PrismaService],
-  exports: [PrismaService],
-})
-export class PrismaModule {}
