+++ backend/src/threat_level_code/threat_level_code.service.spec.ts
@@ -1,13 +1,13 @@
 import { Test, TestingModule } from "@nestjs/testing";
 import { ThreatLevelCodeService } from "./threat_level_code.service";
-import { PrismaModule } from "nestjs-prisma";
+import { PrismaModuleCaseManagement } from "../prisma/cm/prisma.cm.module";
 
 describe("ThreatLevelCodeService", () => {
   let service: ThreatLevelCodeService;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      imports: [PrismaModule],
+      imports: [PrismaModuleCaseManagement],
       providers: [ThreatLevelCodeService],
     }).compile();
